adm: rename `morph` to `fschain`
Leave compatibility with the `morph` command and warn if it executes.

Refs #2702.
