+++ cmd/neofs-adm/internal/modules/fschain/balance.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"crypto/elliptic"
