+++ cmd/neofs-adm/internal/modules/fschain/container.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"encoding/json"
