+++ cmd/neofs-adm/internal/modules/fschain/deploy.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"encoding/json"
