+++ cmd/neofs-adm/internal/modules/fschain/download.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"context"
