+++ cmd/neofs-adm/internal/modules/fschain/dump_hashes.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"bytes"
