+++ cmd/neofs-adm/internal/modules/fschain/generate.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"errors"
