+++ cmd/neofs-adm/internal/modules/fschain/initialize_nns.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"fmt"
