+++ cmd/neofs-adm/internal/modules/fschain/local_client.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"crypto/elliptic"
