+++ cmd/neofs-adm/internal/modules/fschain/n3client.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"context"
