+++ cmd/neofs-adm/internal/modules/fschain/netmap_candidates.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"fmt"
