+++ cmd/neofs-adm/internal/modules/fschain/policy.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"fmt"
