+++ cmd/neofs-adm/internal/modules/fschain/remove_node.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"errors"
