+++ cmd/neofs-adm/internal/modules/fschain/renew_domain.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"errors"
