+++ cmd/neofs-adm/internal/modules/fschain/verified_domains.go
@@ -1,4 +1,4 @@
-package morph
+package fschain
 
 import (
 	"errors"
