+++ cmd/neofs-adm/internal/modules/root.go
@@ -2,9 +2,10 @@ package modules
 
 import (
 	"os"
+	"slices"
 
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-adm/internal/modules/config"
-	"github.com/nspcc-dev/neofs-node/cmd/neofs-adm/internal/modules/morph"
+	"github.com/nspcc-dev/neofs-node/cmd/neofs-adm/internal/modules/fschain"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-adm/internal/modules/storagecfg"
 	"github.com/nspcc-dev/neofs-node/misc"
 	"github.com/nspcc-dev/neofs-node/pkg/util/autocomplete"
@@ -39,11 +40,15 @@ func init() {
 	rootCmd.Flags().Bool("version", false, "Application version")
 
 	rootCmd.AddCommand(config.RootCmd)
-	rootCmd.AddCommand(morph.RootCmd)
+	rootCmd.AddCommand(fschain.RootCmd)
 	rootCmd.AddCommand(storagecfg.RootCmd)
 
 	rootCmd.AddCommand(autocomplete.Command("neofs-adm"))
 	rootCmd.AddCommand(gendoc.Command(rootCmd))
+
+	if slices.Contains(os.Args, "morph") {
+		rootCmd.Println("Warning: 'morph' is deprecated, use 'fschain' instead")
+	}
 }
 
 func Execute() error {
